+++ src/navigations/AuthStack.js
@@ -12,6 +12,7 @@ import TermScreen from '../screens/TermScreen';
 import PrivacyPoliceScreen from '../screens/PrivacyPolicyScreen';
 import CalanderView from '../screens/Calendar/CalendarView';
 import ToDo from '../screens/Calendar/ToDo';
+// import Diary from '../screens/Diary2';
 import Diary3 from '../screens/Diary3';
 import { AuthRoutes } from './routes';
